# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket API",
        "description": "API документация для управления билетами и мероприятиями.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Gateway server"
        }
    ],
    "paths": {
        "/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Список мест события {id}",
                "description": "Возвращает данные о местах события {id}.",
                "operationId": "f2f0a5a610e5bd6a572f3b9bd5dd52c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID места",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "x": {
                                            "description": "Координата X",
                                            "type": "number",
                                            "format": "float",
                                            "example": 0
                                        },
                                        "y": {
                                            "description": "Координата Y",
                                            "type": "number",
                                            "format": "float",
                                            "example": 0
                                        },
                                        "width": {
                                            "description": "Ширина",
                                            "type": "number",
                                            "format": "float",
                                            "example": 20
                                        },
                                        "height": {
                                            "description": "Высота",
                                            "type": "number",
                                            "format": "float",
                                            "example": 20
                                        },
                                        "is_available": {
                                            "description": "Место доступно",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{id}/book": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Резервация места",
                "description": "Позволяет резервировать указанное количество мест.",
                "operationId": "702427869d7b5055c4ab3886d16f32cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID события",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "places": {
                                        "description": "Список ID мест",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "name": {
                                        "description": "Имя покупателя",
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Результат резервации мест и ид_резерва",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Результат бронирования",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "reservation_id": {
                                            "description": "ID резерва",
                                            "type": "string",
                                            "example": "confirmed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/shows": {
            "get": {
                "tags": [
                    "Shows"
                ],
                "summary": "Список мероприятий",
                "description": "Возвращает список доступных мероприятий.",
                "operationId": "e4c02406ed18af63f1eca62aff7aba97",
                "responses": {
                    "200": {
                        "description": "Список мероприятий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID мероприятия",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Название мероприятия",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/shows/{id}": {
            "get": {
                "tags": [
                    "Shows"
                ],
                "summary": "Получение детальной информации по конкретному События",
                "description": "Возвращает подробную информацию о События.",
                "operationId": "7fc9c7472858b74a945811fc160543e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID мероприятия",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детальная страница События",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID события",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "showId": {
                                            "description": "ID мероприятия",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "date": {
                                            "description": "Дата события",
                                            "type": "string",
                                            "example": "2019-08-22 20:26:38"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Events",
            "description": "API Endpoints для просмотра и резерва мест"
        },
        {
            "name": "Shows",
            "description": "API Endpoints для управления мероприятиями"
        }
    ]
}